fix: command not found pacman

+++ system/aliases.zsh
@@ -1,14 +1,23 @@
 alias ..="cd .."
 alias ...="cd ../.."
 alias py="python"
-alias checkup="sudo pamac checkupdates -a"
-alias up="pamac upgrade -a --no-confirm"
-alias upnup="sudo pacman -Syyuu"
-alias cleanup="sudo pacman -Rsn $(pacman -Qdtq)"
-alias buildup="pamac build --no-confirm"
 alias npm="echo 'WTF! YOU SHOULD USE \"yarn\"' && npm"
 alias myip="ip addr show wlo1 | grep wlo1"
-# alias k="microk8s.kubectl"
-# alias kube="microk8s.kubectl"
-# alias kubectl="microk8s.kubectl"
+
+if command -v pamac >/dev/null 2>&1; then
+    alias checkup="sudo pamac checkupdates -a"
+    alias up="pamac upgrade -a --no-confirm"
+    alias buildup="pamac build --no-confirm"
+fi
+
+if command -v pacman >/dev/null 2>&1; then
+    alias upnup="sudo pacman -Syyuu"
+    alias cleanup="sudo pacman -Rsn $(pacman -Qdtq)"
+fi
+
+if command -v microk8s >/dev/null 2>&1; then
+    alias k="microk8s.kubectl"
+    alias kube="microk8s.kubectl"
+    alias kubectl="microk8s.kubectl"
+fi
